refactor(interpreter): split activation dispatch and validation tests
