+++ src/interpreter/src/activation/tests/mod.rs
@@ -1,4 +1,9 @@
     use super::*;
+    mod registers;
+    mod captures;
+    mod guards;
+    mod exhaustiveness;
+    mod dispatch;
     mod registration;
     use std::collections::HashMap;
     use std::sync::{
@@ -135,236 +140,6 @@
             ));
         }
         cases
-    }
-
-    #[test]
-    fn activation_capture_slot_supports_all_enabled_scalar_kinds() {
-        let interpreter = Interpreter::new_with_full_stdlib(0);
-        for (kind, source) in scalar_capture_cases() {
-            let slot = create_capture_slot_for_kind(&kind, &interpreter).unwrap();
-            let cells_before = slot.reactive_root_cell_ids();
-            assert_eq!(cells_before.len(), 1);
-            commit_capture_slot(&slot, &source).unwrap();
-            assert_eq!(slot, source);
-            assert_eq!(slot.reactive_root_cell_ids(), cells_before);
-        }
-    }
-
-    #[cfg(any(feature = "string", feature = "variable_define"))]
-    #[test]
-    fn activation_capture_slot_preserves_identity_across_updates() {
-        let interpreter = Interpreter::new_with_full_stdlib(0);
-        let slot = create_capture_slot_for_kind(&ValueKind::String, &interpreter).unwrap();
-        let cells = slot.reactive_root_cell_ids();
-        commit_capture_slot(&slot, &Value::String(Ref::new("first".to_string()))).unwrap();
-        assert_eq!(slot, Value::String(Ref::new("first".to_string())));
-        assert_eq!(slot.reactive_root_cell_ids(), cells);
-        commit_capture_slot(&slot, &Value::String(Ref::new("second".to_string()))).unwrap();
-        assert_eq!(slot, Value::String(Ref::new("second".to_string())));
-        assert_eq!(slot.reactive_root_cell_ids(), cells);
-    }
-
-    #[cfg(all(
-        feature = "tuple",
-        feature = "enum",
-        feature = "record",
-        feature = "map",
-        feature = "set",
-        feature = "table",
-        feature = "string",
-        feature = "f64"
-    ))]
-    #[test]
-    fn activation_capture_slots_support_enabled_composite_value_kinds() {
-        let interpreter = Interpreter::new_with_full_stdlib(0);
-        let enum_id = hash_str("capture-enum");
-        let variant_id = hash_str("payload");
-        let names = Ref::new(HashMap::from([
-            (enum_id, "capture-enum".to_string()),
-            (variant_id, "payload".to_string()),
-        ]));
-        let cases = vec![
-            Value::Tuple(Ref::new(MechTuple::from_vec(vec![
-                Value::F64(Ref::new(1.0)),
-                Value::String(Ref::new("tuple".to_string())),
-            ]))),
-            Value::Enum(Ref::new(MechEnum {
-                id: enum_id,
-                variants: vec![(variant_id, Some(Value::F64(Ref::new(2.0))))],
-                names,
-            })),
-            Value::Record(Ref::new(MechRecord::new(vec![
-                ("field", Value::F64(Ref::new(3.0))),
-            ]))),
-            Value::Map(Ref::new(MechMap::from_vec(vec![(
-                Value::String(Ref::new("key".to_string())),
-                Value::F64(Ref::new(4.0)),
-            )]))),
-            Value::Set(Ref::new(MechSet::from_vec(vec![Value::String(Ref::new(
-                "member".to_string(),
-            ))]))),
-            Value::Table(Ref::new(MechTable::new_table(
-                vec!["column".to_string()],
-                vec![ValueKind::F64],
-                vec![vec![Value::F64(Ref::new(5.0)), Value::F64(Ref::new(6.0))]],
-            ))),
-        ];
-
-        for source in cases {
-            let kind = source.kind();
-            let slot = create_capture_slot_for_kind(&kind, &interpreter).unwrap();
-            let cells = slot.reactive_root_cell_ids();
-            assert_eq!(cells.len(), 1, "missing stable root for {kind}");
-            commit_capture_slot(&slot, &source).unwrap();
-            assert_eq!(slot, source);
-            assert_eq!(slot.reactive_root_cell_ids(), cells);
-        }
-    }
-
-    #[cfg(all(feature = "f64", feature = "string"))]
-    #[test]
-    fn activation_capture_commit_validates_every_binding_before_mutation() {
-        let interpreter = Interpreter::new_with_full_stdlib(0);
-        let number = ActivationPatternCapture {
-            id: hash_str("number"),
-            name: "number".to_string(),
-            kind: ValueKind::F64,
-            proposed: create_capture_slot_for_kind(&ValueKind::F64, &interpreter).unwrap(),
-            committed: create_capture_slot_for_kind(&ValueKind::F64, &interpreter).unwrap(),
-        };
-        let text = ActivationPatternCapture {
-            id: hash_str("text"),
-            name: "text".to_string(),
-            kind: ValueKind::String,
-            proposed: create_capture_slot_for_kind(&ValueKind::String, &interpreter).unwrap(),
-            committed: create_capture_slot_for_kind(&ValueKind::String, &interpreter).unwrap(),
-        };
-        let captures = vec![number, text];
-        let attempted = PatternMatch {
-            matched: true,
-            bindings: vec![
-                PatternBinding {
-                    index: 0,
-                    id: hash_str("number"),
-                    name: "number".to_string(),
-                    kind: ValueKind::F64,
-                    value: Value::F64(Ref::new(9.0)),
-                },
-                PatternBinding {
-                    index: 1,
-                    id: hash_str("text"),
-                    name: "text".to_string(),
-                    kind: ValueKind::F64,
-                    value: Value::F64(Ref::new(10.0)),
-                },
-            ],
-        };
-
-        let error = ReactiveBindingSink {
-            captures: &captures,
-        }
-        .commit(&attempted)
-        .unwrap_err();
-        assert_eq!(error.kind_name(), "ActivationPatternCaptureKindUnsupported");
-        assert_eq!(captures[0].proposed, Value::F64(Ref::new(0.0)));
-        assert_eq!(captures[1].proposed, Value::String(Ref::new(String::new())));
-        assert_eq!(captures[0].committed, Value::F64(Ref::new(0.0)));
-        assert_eq!(
-            captures[1].committed,
-            Value::String(Ref::new(String::new()))
-        );
-    }
-
-    #[cfg(all(feature = "f64", feature = "string"))]
-    #[test]
-    fn activation_capture_gate_validates_entire_commit_before_mutation_or_pulse() {
-        let captures = vec![
-            ActivationPatternCapture {
-                id: hash_str("number"),
-                name: "number".to_string(),
-                kind: ValueKind::F64,
-                proposed: Value::F64(Ref::new(9.0)),
-                committed: Value::F64(Ref::new(1.0)),
-            },
-            ActivationPatternCapture {
-                id: hash_str("text"),
-                name: "text".to_string(),
-                kind: ValueKind::String,
-                proposed: Value::F64(Ref::new(10.0)),
-                committed: Value::String(Ref::new("before".to_string())),
-            },
-        ];
-        let selected = Ref::new(0);
-        let pulse = Ref::new(0);
-        let gate = Gate {
-            arm: 0,
-            selected,
-            captures: captures.clone(),
-            out: pulse.clone(),
-        };
-
-        let error = gate.solve_reactive().unwrap_err();
-        assert_eq!(error.kind_name(), "ActivationPatternCaptureKindUnsupported");
-        assert_eq!(captures[0].committed, Value::F64(Ref::new(1.0)));
-        assert_eq!(
-            captures[1].committed,
-            Value::String(Ref::new("before".to_string()))
-        );
-        assert_eq!(
-            *pulse.borrow(),
-            0,
-            "body pulse must follow a successful commit"
-        );
-    }
-
-    #[cfg(feature = "atom")]
-    #[test]
-    fn activation_atom_capture_accepts_a_new_atom_value() {
-        let mut interpreter = interpret(
-            r#"
-event := :first
-~> event
-  | captured => {
-      selected := captured
-    }
-  | * => {
-      selected := :fallback
-    }
-"#,
-        );
-        let trigger = root_cell(&interpreter, "event");
-        let topology = plan_snapshot(&interpreter);
-        let registration = registration(&interpreter);
-        let Value::Atom(event) = symbol(&interpreter, "event") else {
-            panic!("event is not an atom")
-        };
-        *event.borrow_mut() = MechAtom::from_name("second");
-
-        let outcome = interpreter.advance_reactive_turn(&[trigger]).unwrap();
-        assert_eq!(selected_arm_index(&registration, &outcome), 0);
-        let selected_atom = {
-            let plan = interpreter.plan();
-            let plan = plan.borrow();
-            (registration.arms[0].body_node_start..registration.arms[0].body_node_end)
-                .rev()
-                .find_map(|node| match detached(&plan.node(node).unwrap().function.out()) {
-                    Value::Atom(atom) => Some(atom.borrow().id()),
-                    _ => None,
-                })
-                .expect("no atom output in selected arm body")
-        };
-        assert_eq!(selected_atom, hash_str("second"));
-        assert_eq!(plan_snapshot(&interpreter), topology);
-    }
-
-    #[cfg(all(feature = "f64", any(feature = "string", feature = "variable_define")))]
-    #[test]
-    fn activation_capture_slot_rejects_kind_mismatch() {
-        let interpreter = Interpreter::new_with_full_stdlib(0);
-        let slot = create_capture_slot_for_kind(&ValueKind::F64, &interpreter).unwrap();
-        let error =
-            commit_capture_slot(&slot, &Value::String(Ref::new("wrong".to_string()))).unwrap_err();
-        assert_eq!(error.kind_name(), "ActivationPatternCaptureKindUnsupported");
     }
 
     type PlanSnapshot = (
@@ -705,423 +480,6 @@ event := :first
     }
 
     #[test]
-    fn activation_final_binding_is_exhaustive_and_guarded_binding_falls_through_to_it() {
-        let mut interpreter = interpret(
-            r#"
-physics-tick := 0.25
-~position-y := -1.0
-~velocity-y := 0.0
-floor := 0.0
-restitution := 0.5
-gravity := 4.0
-
-~> physics-tick
-  | dt, position-y >= floor => {
-      velocity-y = -velocity-y * restitution
-      position-y = floor
-    }
-  | dt => {
-      velocity-y = velocity-y + gravity * dt
-      position-y = position-y + velocity-y * dt
-    }
-"#,
-        );
-        let trigger = root_cell(&interpreter, "physics-tick");
-        let activation = registration(&interpreter);
-        let topology = plan_snapshot(&interpreter);
-        assert_eq!(activation.arms.len(), 2);
-        assert!(activation.arms[0].guard.is_some());
-        assert!(activation.arms[1].guard.is_none());
-        assert_eq!(activation.arms[0].captures.len(), 1);
-        assert_eq!(activation.arms[1].captures.len(), 1);
-        let arm_registers = (0..activation.arms.len())
-            .map(|arm| arm_register_nodes(&interpreter, &activation, arm))
-            .collect::<Vec<_>>();
-        assert!(arm_registers.iter().all(|registers| registers.len() == 2));
-        assert_eq!(
-            (
-                f64_symbol(&interpreter, "position-y"),
-                f64_symbol(&interpreter, "velocity-y"),
-            ),
-            (-1.0, 0.0),
-        );
-
-        let fallback = interpreter.advance_reactive_turn(&[trigger]).unwrap();
-        assert_eq!(selected_arm_index(&activation, &fallback), 1);
-        assert_eq!(fallback.register_commit.committed_nodes, arm_registers[1]);
-        assert_eq!(
-            (
-                f64_symbol(&interpreter, "position-y"),
-                f64_symbol(&interpreter, "velocity-y"),
-            ),
-            (-1.0, 1.0),
-        );
-        assert_eq!(
-            committed_capture_value(&interpreter, 1, 0),
-            Value::F64(Ref::new(0.25)),
-        );
-        assert_eq!(plan_snapshot(&interpreter), topology);
-
-        set_f64_symbol(&interpreter, "position-y", 1.0);
-        let guarded = interpreter.advance_reactive_turn(&[trigger]).unwrap();
-        assert_eq!(selected_arm_index(&activation, &guarded), 0);
-        assert_eq!(guarded.register_commit.committed_nodes, arm_registers[0]);
-        assert_eq!(
-            (
-                f64_symbol(&interpreter, "position-y"),
-                f64_symbol(&interpreter, "velocity-y"),
-            ),
-            (0.0, -0.5),
-        );
-        assert_eq!(
-            committed_capture_value(&interpreter, 0, 0),
-            Value::F64(Ref::new(0.25)),
-        );
-        assert_eq!(plan_snapshot(&interpreter), topology);
-    }
-
-    #[test]
-    fn activation_guarded_wildcard_may_precede_an_exhaustive_binding() {
-        let mut interpreter = interpret(
-            r#"
-event := 2.0
-ready := false
-
-~> event
-  | *, ready == true => {
-      selected := -1.0
-    }
-  | value => {
-      selected := value
-    }
-"#,
-        );
-        let trigger = root_cell(&interpreter, "event");
-        let topology = plan_snapshot(&interpreter);
-        let outcome = interpreter.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&interpreter, &topology, &outcome, 1, 2.0);
-    }
-
-    #[test]
-    fn activation_final_irrefutable_tuple_is_exhaustive() {
-        let mut interpreter = interpret(
-            r#"
-event := (1.0, 2.0)
-~> event
-  | (tick, dt) => {
-      selected := tick * 10.0 + dt
-    }
-"#,
-        );
-        let trigger = root_cell(&interpreter, "event");
-        let topology = plan_snapshot(&interpreter);
-        set_tuple_event(
-            &interpreter,
-            vec![Value::F64(Ref::new(3.0)), Value::F64(Ref::new(4.0))],
-        );
-        let outcome = interpreter.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&interpreter, &topology, &outcome, 0, 34.0);
-    }
-
-    #[cfg(all(feature = "matrix", feature = "f64"))]
-    #[test]
-    fn activation_final_irrefutable_fixed_matrix_pattern_is_exhaustive() {
-        let mut interpreter = interpret(
-            r#"
-event := [1.0 2.0]
-~> event
-  | [left, right] => {
-      selected := left * 10.0 + right
-    }
-"#,
-        );
-        let trigger = root_cell(&interpreter, "event");
-        let topology = plan_snapshot(&interpreter);
-        set_f64_matrix_event(&interpreter, vec![3.0, 4.0]);
-        let outcome = interpreter.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&interpreter, &topology, &outcome, 0, 34.0);
-    }
-
-    #[test]
-    fn activation_final_refutable_pattern_is_non_exhaustive() {
-        for pattern in ["1.0", "(x, x)"] {
-            let mut interpreter = if pattern == "1.0" {
-                interpret("event := 1.0")
-            } else {
-                interpret("event := (1.0, 1.0)")
-            };
-            let topology = plan_snapshot(&interpreter);
-            let error = interpret_more(
-                &mut interpreter,
-                &format!(
-                    "~> event\n  | {pattern} => {{\n      selected := 1.0\n    }}"
-                ),
-            )
-            .unwrap_err();
-            assert_eq!(error.kind_name(), "ActivationPatternArmsNonExhaustive");
-            assert_eq!(plan_snapshot(&interpreter), topology);
-        }
-
-        let mut interpreter = interpret("event := 1.0");
-        let topology = plan_snapshot(&interpreter);
-        let error = interpret_more(
-            &mut interpreter,
-            r#"
-~> event
-  | value, value > 0.0 => {
-      selected := value
-    }
-"#,
-        )
-        .unwrap_err();
-        assert_eq!(error.kind_name(), "ActivationPatternArmsNonExhaustive");
-        assert_eq!(plan_snapshot(&interpreter), topology);
-    }
-
-    #[test]
-    fn activation_unguarded_wildcard_before_the_final_arm_is_rejected() {
-        let mut interpreter = interpret("event := 1.0");
-        let topology = plan_snapshot(&interpreter);
-        let error = interpret_more(
-            &mut interpreter,
-            r#"
-~> event
-  | * => {
-      selected := 1.0
-    }
-  | value => {
-      selected := value
-    }
-"#,
-        )
-        .unwrap_err();
-        assert_eq!(error.kind_name(), "ActivationPatternWildcardMustBeLast");
-        assert_eq!(plan_snapshot(&interpreter), topology);
-    }
-
-    #[test]
-    fn activation_selected_arm_commits_register_batch_and_other_arms_schedule_nothing() {
-        let mut interpreter = interpret(
-            r#"
-event := 0.0
-~x := 1.0
-~y := 10.0
-
-~> event
-  | first, first > 10.0 => {
-      x = first
-      y += first
-    }
-  | second, second > 0.0 => {
-      x = second * 2.0
-      y = second + 1.0
-    }
-  | * => {
-      x = -1.0
-      y = -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&interpreter, "event");
-        let activation = registration(&interpreter);
-        let topology = plan_snapshot(&interpreter);
-        let arm_registers = (0..activation.arms.len())
-            .map(|arm| arm_register_nodes(&interpreter, &activation, arm))
-            .collect::<Vec<_>>();
-        assert!(arm_registers.iter().all(|registers| registers.len() == 2));
-        {
-            let plan = interpreter.plan();
-            let plan = plan.borrow();
-            for (arm_index, registers) in arm_registers.iter().enumerate() {
-                for register in registers {
-                    let node = plan.node(*register).unwrap();
-                    assert!(node.inputs.iter().any(|dependency| {
-                        dependency.cell == activation.arms[arm_index].pulse_cell
-                            && dependency.kind == ReactiveDependencyKind::Reactive
-                    }));
-                    for (other_arm, registration) in activation.arms.iter().enumerate() {
-                        if other_arm != arm_index {
-                            assert!(!node.inputs.iter().any(|dependency| {
-                                dependency.cell == registration.pulse_cell
-                                    && dependency.kind == ReactiveDependencyKind::Reactive
-                            }));
-                        }
-                    }
-                    for capture in &activation.arms[arm_index].captures {
-                        assert!(node.inputs.iter().any(|dependency| {
-                            dependency.cell == capture.cell
-                                && dependency.kind == ReactiveDependencyKind::Sampled
-                        }));
-                    }
-                }
-            }
-        }
-
-        // Initial dispatch selects the fallback only to seed captures. No arm
-        // body is pulsed while the static graph is being registered.
-        assert_eq!(
-            (f64_symbol(&interpreter, "x"), f64_symbol(&interpreter, "y")),
-            (1.0, 10.0)
-        );
-        assert!(!interpreter.has_pending_reactive_registers());
-        for arm in 0..activation.arms.len() {
-            assert_eq!(arm_pulse_generation(&interpreter, arm), 0);
-        }
-
-        // Both guarded arms are eligible, so source order selects arm zero.
-        // The matching-but-losing arm must not schedule either register.
-        set_f64_symbol(&interpreter, "event", 20.0);
-        let first = interpreter.advance_reactive_turn(&[trigger]).unwrap();
-        assert_eq!(selected_arm_index(&activation, &first), 0);
-        assert_eq!(first.before_commit.pending_register_nodes, arm_registers[0]);
-        assert_eq!(first.register_commit.staged_nodes, arm_registers[0]);
-        assert_eq!(first.register_commit.committed_nodes, arm_registers[0]);
-        assert_eq!(
-            (f64_symbol(&interpreter, "x"), f64_symbol(&interpreter, "y")),
-            (20.0, 30.0)
-        );
-        for register in arm_registers[1].iter().chain(&arm_registers[2]) {
-            assert!(!first.before_commit.pending_register_nodes.contains(register));
-            assert!(!first.register_commit.staged_nodes.contains(register));
-            assert!(!first.register_commit.committed_nodes.contains(register));
-            assert!(!first.after_commit.pending_register_nodes.contains(register));
-        }
-
-        // Arm zero still matches, but its false guard must leave its register
-        // nodes dormant. Arm one's capture is consumed by both writes from
-        // this same trigger.
-        set_f64_symbol(&interpreter, "event", 5.0);
-        let second = interpreter.advance_reactive_turn(&[trigger]).unwrap();
-        assert_eq!(selected_arm_index(&activation, &second), 1);
-        assert_eq!(second.before_commit.pending_register_nodes, arm_registers[1]);
-        assert_eq!(second.register_commit.staged_nodes, arm_registers[1]);
-        assert_eq!(second.register_commit.committed_nodes, arm_registers[1]);
-        assert_eq!(
-            (f64_symbol(&interpreter, "x"), f64_symbol(&interpreter, "y")),
-            (10.0, 6.0)
-        );
-        assert_eq!(
-            committed_capture_value(&interpreter, 1, 0),
-            Value::F64(Ref::new(5.0))
-        );
-        for register in arm_registers[0].iter().chain(&arm_registers[2]) {
-            assert!(!second.before_commit.pending_register_nodes.contains(register));
-            assert!(!second.register_commit.staged_nodes.contains(register));
-            assert!(!second.register_commit.committed_nodes.contains(register));
-            assert!(!second.after_commit.pending_register_nodes.contains(register));
-        }
-        assert!(!interpreter.has_pending_reactive_registers());
-        assert_eq!(plan_snapshot(&interpreter), topology);
-    }
-
-    #[test]
-    fn activation_equal_trigger_packets_produce_distinct_register_transitions_without_plan_growth() {
-        let mut interpreter = interpret(
-            r#"
-event := 2.0
-~count := 0.0
-
-~> event
-  | amount => {
-      count += amount
-    }
-  | * => {
-      fallback := 0.0
-    }
-"#,
-        );
-        let trigger = root_cell(&interpreter, "event");
-        let activation = registration(&interpreter);
-        let registers = arm_register_nodes(&interpreter, &activation, 0);
-        let topology = plan_snapshot(&interpreter);
-        assert_eq!(registers.len(), 1);
-        assert_eq!(f64_symbol(&interpreter, "count"), 0.0);
-
-        for (expected_count, expected_pulse) in [(2.0, 1usize), (4.0, 2usize)] {
-            let outcome = interpreter.advance_reactive_turn(&[trigger]).unwrap();
-            assert_eq!(selected_arm_index(&activation, &outcome), 0);
-            assert_eq!(outcome.before_commit.pending_register_nodes, registers);
-            assert_eq!(outcome.register_commit.staged_nodes, registers);
-            assert_eq!(outcome.register_commit.committed_nodes, registers);
-            assert_eq!(f64_symbol(&interpreter, "count"), expected_count);
-            assert_eq!(arm_pulse_generation(&interpreter, 0), expected_pulse);
-            assert_eq!(plan_snapshot(&interpreter), topology);
-        }
-    }
-
-    #[test]
-    fn activation_arm_alias_of_live_input_remains_sampled_until_trigger() {
-        let mut interpreter = Interpreter::new_with_full_stdlib(0);
-        let outer_id = hash_str("outer");
-        {
-            let symbols = interpreter.symbols();
-            let mut symbols = symbols.borrow_mut();
-            symbols.insert(outer_id, Value::F64(Ref::new(1.0)), true);
-            symbols
-                .dictionary
-                .borrow_mut()
-                .insert(outer_id, "outer".to_string());
-        }
-        interpreter
-            .dictionary()
-            .borrow_mut()
-            .insert(outer_id, "outer".to_string());
-        interpret_more(
-            &mut interpreter,
-            r#"
-event := 0.0
-~state := 0.0
-
-~> event
-  | tick => {
-      sampled := outer
-      state = sampled
-    }
-"#,
-        )
-        .unwrap();
-
-        let trigger = root_cell(&interpreter, "event");
-        let outer = root_cell(&interpreter, "outer");
-        let activation = registration(&interpreter);
-        let registers = arm_register_nodes(&interpreter, &activation, 0);
-        let topology = plan_snapshot(&interpreter);
-        assert_eq!(registers.len(), 1);
-        assert_eq!(f64_symbol(&interpreter, "state"), 0.0);
-        {
-            let plan = interpreter.plan();
-            let plan = plan.borrow();
-            let register = plan.node(registers[0]).unwrap();
-            assert!(register.inputs.iter().any(|dependency| {
-                dependency.cell == outer
-                    && dependency.kind == ReactiveDependencyKind::Sampled
-            }));
-            assert!(register.inputs.iter().any(|dependency| {
-                dependency.cell == activation.arms[0].pulse_cell
-                    && dependency.kind == ReactiveDependencyKind::Reactive
-            }));
-        }
-
-        set_f64_symbol(&interpreter, "outer", 5.0);
-        let sampled_only = interpreter.advance_reactive_turn(&[outer]).unwrap();
-        assert!(!sampled_only
-            .before_commit
-            .pending_register_nodes
-            .contains(&registers[0]));
-        assert!(!sampled_only
-            .register_commit
-            .committed_nodes
-            .contains(&registers[0]));
-        assert_eq!(f64_symbol(&interpreter, "state"), 0.0);
-
-        let tick = interpreter.advance_reactive_turn(&[trigger]).unwrap();
-        assert_eq!(selected_arm_index(&activation, &tick), 0);
-        assert_eq!(tick.register_commit.committed_nodes, registers);
-        assert_eq!(f64_symbol(&interpreter, "state"), 5.0);
-        assert_eq!(plan_snapshot(&interpreter), topology);
-    }
-
-    #[test]
     fn activation_failed_multi_register_staging_mutates_nothing() {
         let mut interpreter = interpret(
             r#"
@@ -1246,261 +604,6 @@ event<event-kind> := :pressed(0.0)
         (interpreter, trigger, registration, topology)
     }
 
-    #[test]
-    fn activation_pattern_selects_pressed_released_and_wildcard() {
-        let (mut i, trigger, _, topology) = load_enum_activation();
-        for (name, payload, arm, output) in [
-            ("pressed", 10., 0, 10.),
-            ("released", 20., 1, 1020.),
-            ("other", 30., 2, -1.),
-        ] {
-            set_enum_event(&i, name, payload);
-            let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-            assert_dispatch_turn(&i, &topology, &outcome, arm, output);
-        }
-    }
-    #[test]
-    fn activation_pattern_enum_arms_compile_independent_of_initial_variant() {
-        let (mut i, trigger, r, topology) = load_enum_activation();
-        assert_eq!(r.arms[1].captures[0].kind, ValueKind::F64);
-        set_enum_event(&i, "released", 20.);
-        let o = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &o, 1, 1020.);
-    }
-    #[test]
-    fn activation_pattern_enum_payload_capture_is_available() {
-        let (mut i, trigger, r, topology) = load_enum_activation();
-        let cell = r.arms[0].captures[0].cell;
-        assert!(
-            i.plan().borrow().nodes[r.arms[0].body_node_start..r.arms[0].body_node_end]
-                .iter()
-                .any(|n| n.inputs.iter().any(|d| d.cell == cell))
-        );
-        set_enum_event(&i, "pressed", 10.);
-        let o = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &o, 0, 10.);
-    }
-    #[test]
-    fn activation_pattern_equal_packets_dispatch_repeatedly() {
-        let (mut i, trigger, _, topology) = load_enum_activation();
-        set_enum_event(&i, "pressed", 30.);
-        for _ in 0..2 {
-            let o = i.advance_reactive_turn(&[trigger]).unwrap();
-            assert_dispatch_turn(&i, &topology, &o, 0, 30.);
-        }
-    }
-    #[test]
-    fn activation_pattern_unselected_arm_nodes_do_not_execute() {
-        let (mut i, trigger, r, topology) = load_enum_activation();
-        set_enum_event(&i, "released", 20.);
-        let o = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &o, 1, 1020.);
-        let executed = turn_executed_nodes(&o);
-        for arm in [&r.arms[0], &r.arms[2]] {
-            for node in arm.body_node_start..arm.body_node_end {
-                assert!(!executed.contains(&node));
-            }
-        }
-    }
-    #[test]
-    fn activation_only_selected_arm_commits_matching_captures() {
-        let mut i = interpret(
-            r#"
-event := 1.0
-~> event
-  | first => {
-      selected := first
-    }
-  | later => {
-      rejected := later
-    }
-  | * => {
-      fallback := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let activation = registration(&i);
-        let later_before = committed_capture_value(&i, 1, 0);
-        let committed_cell = activation.arms[0].captures[0].cell;
-        let proposed_cell = proposed_capture_value(&i, 0, 0).reactive_root_cell_ids()[0];
-        assert_ne!(proposed_cell, committed_cell);
-        {
-            let plan = i.plan();
-            let plan = plan.borrow();
-            assert!(
-                plan.nodes[activation.arms[0].matcher_node]
-                    .outputs
-                    .contains(&proposed_cell)
-            );
-            assert!(
-                !plan.nodes[activation.arms[0].matcher_node]
-                    .outputs
-                    .contains(&committed_cell)
-            );
-            assert!(
-                plan.nodes[activation.arms[0].gate_node]
-                    .outputs
-                    .contains(&committed_cell)
-            );
-            let body_inputs = plan.nodes
-                [activation.arms[0].body_node_start..activation.arms[0].body_node_end]
-                .iter()
-                .flat_map(|node| node.inputs.iter().map(|dependency| dependency.cell))
-                .collect::<Vec<_>>();
-            assert!(body_inputs.contains(&committed_cell));
-            assert!(!body_inputs.contains(&proposed_cell));
-        }
-        let Value::F64(event) = symbol(&i, "event") else {
-            panic!("event is not f64")
-        };
-        *event.borrow_mut() = 5.0;
-
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-
-        assert_eq!(selected_arm_index(&activation, &outcome), 0);
-        assert_eq!(committed_capture_value(&i, 0, 0), Value::F64(Ref::new(5.0)));
-        assert_eq!(proposed_capture_value(&i, 1, 0), Value::F64(Ref::new(5.0)));
-        assert_eq!(committed_capture_value(&i, 1, 0), later_before);
-        let executed = turn_executed_nodes(&outcome);
-        for node in activation.arms[1].body_node_start..activation.arms[1].body_node_end {
-            assert!(!executed.contains(&node));
-        }
-    }
-    #[test]
-    fn activation_failed_repeated_binding_leaves_proposed_and_committed_unchanged() {
-        let mut i = interpret(
-            r#"
-event := (1.0, 1.0)
-~> event
-  | (x, x) => {
-      selected := x
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let activation = registration(&i);
-        let proposed_before = proposed_capture_value(&i, 0, 0);
-        let committed_before = committed_capture_value(&i, 0, 0);
-        set_tuple_event(
-            &i,
-            vec![Value::F64(Ref::new(2.0)), Value::F64(Ref::new(3.0))],
-        );
-
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-
-        assert_eq!(selected_arm_index(&activation, &outcome), 1);
-        assert_eq!(proposed_capture_value(&i, 0, 0), proposed_before);
-        assert_eq!(committed_capture_value(&i, 0, 0), committed_before);
-        let executed = turn_executed_nodes(&outcome);
-        for node in activation.arms[0].body_node_start..activation.arms[0].body_node_end {
-            assert!(!executed.contains(&node));
-        }
-    }
-    #[test]
-    fn activation_non_selected_composite_capture_keeps_last_committed_value() {
-        let mut i = interpret(
-            r#"
-event := (0.0, 1.0)
-~> event
-  | (1.0, x) => {
-      selected := x
-    }
-  | later => {
-      rejected := later.2
-    }
-  | * => {
-      fallback := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let activation = registration(&i);
-        let committed_before = committed_capture_value(&i, 1, 0);
-        set_tuple_event(
-            &i,
-            vec![Value::F64(Ref::new(1.0)), Value::F64(Ref::new(10.0))],
-        );
-
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-
-        assert_eq!(selected_arm_index(&activation, &outcome), 0);
-        assert_eq!(committed_capture_value(&i, 1, 0), committed_before);
-        assert_eq!(
-            proposed_capture_value(&i, 1, 0),
-            Value::Tuple(Ref::new(MechTuple::from_vec(vec![
-                Value::F64(Ref::new(1.0)),
-                Value::F64(Ref::new(10.0)),
-            ])))
-        );
-        let executed = turn_executed_nodes(&outcome);
-        for node in activation.arms[1].body_node_start..activation.arms[1].body_node_end {
-            assert!(!executed.contains(&node));
-        }
-    }
-    #[test]
-    fn activation_pattern_switching_arms_does_not_grow_plan() {
-        let (mut i, trigger, _, topology) = load_enum_activation();
-        for (name, payload) in [
-            ("pressed", 10.),
-            ("released", 20.),
-            ("other", 30.),
-            ("pressed", 30.),
-            ("pressed", 30.),
-        ] {
-            set_enum_event(&i, name, payload);
-            i.advance_reactive_turn(&[trigger]).unwrap();
-            assert_eq!(plan_snapshot(&i), topology);
-        }
-    }
-
-    #[test]
-    fn activation_pattern_matches_payload_free_enum_variant() {
-        let mut i = interpret(
-            r#"
-<signal> := :ready | :other
-event<signal> := :other
-~> event
-  | :ready => {
-      selected := 1.0
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let topology = plan_snapshot(&i);
-        set_unit_enum_event(&i, "ready");
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 0, 1.0);
-    }
-    #[test]
-    fn activation_pattern_capture_storage_identity_is_stable() {
-        let (mut i, trigger, r, topology) = load_enum_activation();
-        let captures = r
-            .arms
-            .iter()
-            .flat_map(|arm| arm.captures.iter())
-            .map(|capture| (capture.id, capture.kind.clone(), capture.cell))
-            .collect::<Vec<_>>();
-        for (name, payload) in [("pressed", 10.), ("released", 20.), ("other", 30.)] {
-            set_enum_event(&i, name, payload);
-            i.advance_reactive_turn(&[trigger]).unwrap();
-            let current = registration(&i)
-                .arms
-                .iter()
-                .flat_map(|arm| arm.captures.iter())
-                .map(|capture| (capture.id, capture.kind.clone(), capture.cell))
-                .collect::<Vec<_>>();
-            assert_eq!(current, captures);
-            assert_eq!(plan_snapshot(&i), topology);
-        }
-    }
-
     const ATOM_TUPLE_ACTIVATION: &str = r#"
 event := (:pressed, 0.0)
 ~> event
@@ -1526,43 +629,6 @@ event := (:pressed, 0.0)
         let topology = plan_snapshot(&i);
         (i, trigger, r, topology)
     }
-    #[test]
-    fn activation_pattern_atom_tagged_tuple_selects_arm() {
-        let (mut i, trigger, _, topology) = load_atom_tuple_activation();
-        for (tag, payload, arm, output) in [
-            ("pressed", 10., 0, 10.),
-            ("released", 20., 1, 1020.),
-            ("other", 30., 2, -1.),
-        ] {
-            set_atom_tuple_event(&i, tag, payload);
-            let o = i.advance_reactive_turn(&[trigger]).unwrap();
-            assert_dispatch_turn(&i, &topology, &o, arm, output);
-        }
-    }
-    #[test]
-    fn activation_pattern_atom_tagged_tuple_captures_payload() {
-        let (mut i, trigger, r, topology) = load_atom_tuple_activation();
-        assert_eq!(r.arms[0].captures[0].kind, ValueKind::F64);
-        let cell = r.arms[0].captures[0].cell;
-        assert!(
-            i.plan().borrow().nodes[r.arms[0].body_node_start..r.arms[0].body_node_end]
-                .iter()
-                .any(|n| n.inputs.iter().any(|d| d.cell == cell))
-        );
-        set_atom_tuple_event(&i, "pressed", 10.);
-        let o = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &o, 0, 10.);
-    }
-    #[test]
-    fn activation_pattern_atom_tuple_arms_compile_independent_of_initial_tag() {
-        let (mut i, trigger, r, topology) = load_atom_tuple_activation();
-        assert_eq!(r.arms[0].captures[0].kind, ValueKind::F64);
-        assert_eq!(r.arms[1].captures[0].kind, ValueKind::F64);
-        set_atom_tuple_event(&i, "released", 20.);
-        let o = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &o, 1, 1020.);
-    }
-
     const FLAT_TUPLE_ACTIVATION: &str = r#"
 event := (1.0, 2.0)
 ~> event
@@ -1600,243 +666,6 @@ event := (1.0, 1.0)
         (i, trigger, topology)
     }
     #[test]
-    fn activation_pattern_tuple_captures_elements() {
-        let (mut i, trigger, topology) = tuple_fixture(FLAT_TUPLE_ACTIVATION);
-        set_tuple_event(&i, vec![Value::F64(Ref::new(3.)), Value::F64(Ref::new(4.))]);
-        let o = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &o, 0, 34.);
-    }
-    #[test]
-    fn activation_pattern_nested_tuple_captures_elements() {
-        let (mut i, trigger, topology) = tuple_fixture(NESTED_TUPLE_ACTIVATION);
-        set_tuple_event(
-            &i,
-            vec![
-                Value::Tuple(Ref::new(MechTuple::from_vec(vec![
-                    Value::F64(Ref::new(4.)),
-                    Value::F64(Ref::new(5.)),
-                ]))),
-                Value::F64(Ref::new(6.)),
-            ],
-        );
-        let o = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &o, 0, 456.);
-    }
-    #[test]
-    fn activation_pattern_repeated_capture_requires_equal_values() {
-        let (mut i, trigger, topology) = tuple_fixture(REPEATED_CAPTURE_ACTIVATION);
-        set_tuple_event(&i, vec![Value::F64(Ref::new(2.)), Value::F64(Ref::new(2.))]);
-        let o = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &o, 0, 2.);
-        set_tuple_event(&i, vec![Value::F64(Ref::new(2.)), Value::F64(Ref::new(3.))]);
-        let o = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &o, 1, -1.);
-    }
-
-    #[test]
-    fn activation_guards_fall_through_in_source_order_and_commit_only_the_selected_arm() {
-        let mut i = interpret(
-            r#"
-event := 0.0
-~> event
-  | first, first > 10.0 => {
-      selected := first + 100.0
-    }
-  | second, second > 0.0 => {
-      selected := second + 200.0
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let topology = plan_snapshot(&i);
-        let activation = registration(&i);
-        assert!(activation.arms[0].guard.is_some());
-        assert!(activation.arms[1].guard.is_some());
-        assert!(activation.arms[2].guard.is_none());
-
-        let Value::F64(event) = symbol(&i, "event") else {
-            panic!("event is not f64")
-        };
-        *event.borrow_mut() = 20.0;
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 0, 120.0);
-        let changed = turn_changed_nodes(&outcome);
-        let unchanged = turn_unchanged_nodes(&outcome);
-        for arm in &activation.arms[..2] {
-            let guard = arm.guard.as_ref().unwrap();
-            assert!(changed.contains(&guard.match_gate_node));
-            assert!(changed.contains(&guard.guard_finalizer_node));
-            assert!(unchanged.contains(&arm.finalizer_node));
-        }
-        assert_eq!(committed_capture_value(&i, 0, 0), Value::F64(Ref::new(20.0)));
-        assert_eq!(committed_capture_value(&i, 1, 0), Value::F64(Ref::new(0.0)));
-
-        *event.borrow_mut() = 5.0;
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 1, 205.0);
-        assert_eq!(committed_capture_value(&i, 0, 0), Value::F64(Ref::new(20.0)));
-        assert_eq!(committed_capture_value(&i, 1, 0), Value::F64(Ref::new(5.0)));
-
-        *event.borrow_mut() = -5.0;
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 2, -1.0);
-        assert_eq!(proposed_capture_value(&i, 0, 0), Value::F64(Ref::new(-5.0)));
-        assert_eq!(proposed_capture_value(&i, 1, 0), Value::F64(Ref::new(-5.0)));
-        assert_eq!(committed_capture_value(&i, 0, 0), Value::F64(Ref::new(20.0)));
-        assert_eq!(committed_capture_value(&i, 1, 0), Value::F64(Ref::new(5.0)));
-    }
-
-    #[test]
-    fn activation_guard_outer_dependencies_are_sampled_until_the_next_trigger() {
-        let mut i = interpret(
-            r#"
-event := (:pressed, 0.0)
-threshold := 10.0
-~> event
-  | :pressed(x), x > threshold => {
-      selected := x + 0.0
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let threshold_cell = root_cell(&i, "threshold");
-        let topology = plan_snapshot(&i);
-        let activation = registration(&i);
-        let guard = activation.arms[0].guard.as_ref().unwrap();
-        let Value::F64(threshold) = symbol(&i, "threshold") else {
-            panic!("threshold is not f64")
-        };
-        *threshold.borrow_mut() = 3.0;
-
-        let dependency_turn = i.advance_reactive_turn(&[threshold_cell]).unwrap();
-        let executed = turn_executed_nodes(&dependency_turn);
-        assert!(!executed.contains(&activation.scope_pulse_node));
-        assert!(!executed.contains(&guard.match_gate_node));
-        assert!(!executed.contains(&guard.guard_finalizer_node));
-        for node in guard.guard_node_start..guard.guard_node_end {
-            assert!(!executed.contains(&node));
-        }
-        assert_eq!(plan_snapshot(&i), topology);
-
-        set_atom_tuple_event(&i, "pressed", 5.0);
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 0, 5.0);
-    }
-
-    #[test]
-    fn activation_guard_user_function_refreshes_on_each_matching_trigger() {
-        let mut i = interpret(
-            r#"
-passes(value<f64>, limit<f64>) => <bool>
-  | value > limit.
-
-event := (:pressed, 0.0)
-threshold := 5.0
-~> event
-  | :pressed(x), passes(x, threshold) => {
-      selected := x + 0.0
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let threshold_cell = root_cell(&i, "threshold");
-        let topology = plan_snapshot(&i);
-        {
-            let functions = i.functions();
-            let functions = functions.borrow();
-            let passes = functions.user_functions.get(&hash_str("passes")).unwrap();
-            assert_eq!(passes.code.match_arms.len(), 1);
-            assert!(matches!(passes.code.match_arms[0].pattern, Pattern::Wildcard));
-        }
-
-        set_atom_tuple_event(&i, "pressed", 6.0);
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 0, 6.0);
-
-        set_atom_tuple_event(&i, "pressed", 4.0);
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 1, -1.0);
-
-        let Value::F64(threshold) = symbol(&i, "threshold") else {
-            panic!("threshold is not f64")
-        };
-        *threshold.borrow_mut() = 3.0;
-        let dependency_turn = i.advance_reactive_turn(&[threshold_cell]).unwrap();
-        assert!(turn_executed_nodes(&dependency_turn).is_empty());
-
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 0, 4.0);
-    }
-
-    #[test]
-    fn activation_guard_initialization_commits_the_first_eligible_arm_without_pulsing_a_body() {
-        let i = interpret(
-            r#"
-event := 5.0
-~> event
-  | first, first > 10.0 => {
-      selected := first + 100.0
-    }
-  | second, second > 0.0 => {
-      selected := second + 200.0
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-
-        assert_eq!(proposed_capture_value(&i, 0, 0), Value::F64(Ref::new(5.0)));
-        assert_eq!(proposed_capture_value(&i, 1, 0), Value::F64(Ref::new(5.0)));
-        assert_eq!(committed_capture_value(&i, 0, 0), Value::F64(Ref::new(0.0)));
-        assert_eq!(committed_capture_value(&i, 1, 0), Value::F64(Ref::new(5.0)));
-        assert_eq!(body_output_f64(&i, 1), 205.0);
-        for arm in 0..3 {
-            assert_eq!(arm_pulse_generation(&i, arm), 0);
-        }
-    }
-
-    #[test]
-    fn activation_guard_equal_packets_dispatch_again_without_changing_topology() {
-        let mut i = interpret(
-            r#"
-event := 20.0
-~> event
-  | value, value > 10.0 => {
-      selected := value + 1.0
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let topology = plan_snapshot(&i);
-        for expected_pulse in [1usize, 2] {
-            let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-            assert_dispatch_turn(&i, &topology, &outcome, 0, 21.0);
-            let guard = registration(&i).arms[0].guard.clone().unwrap();
-            let executed = turn_executed_nodes(&outcome);
-            for guard_node in guard.guard_node_start..guard.guard_node_end {
-                assert_eq!(
-                    executed.iter().filter(|node| **node == guard_node).count(),
-                    1
-                );
-            }
-            assert_eq!(arm_pulse_generation(&i, 0), expected_pulse);
-        }
-    }
-
-    #[test]
     fn activation_guard_non_boolean_result_rolls_back_plan_and_symbols() {
         let mut i = interpret("event := 1.0\nouter := 9.0");
         let symbols = i.symbols().borrow().snapshot();
@@ -1862,97 +691,6 @@ event := 20.0
         assert_eq!(*i.dictionary().borrow(), dictionary);
         assert_eq!(plan_snapshot(&i), topology);
         assert_eq!(i.plan().activation_registration_depth(), 0);
-    }
-
-    #[test]
-    fn activation_guard_rejects_unclassified_native_compiler_before_compile() {
-        let mut i = interpret("event := (:released, 1.0)");
-        let compile_calls = Arc::new(AtomicUsize::new(0));
-        i.functions().borrow_mut().insert_function_compiler(
-            "test/eager-guard",
-            Arc::new(EagerGuardTestCompiler {
-                compile_calls: compile_calls.clone(),
-            }),
-        );
-        let symbols = i.symbols().borrow().snapshot();
-        let topology = plan_snapshot(&i);
-
-        let error = interpret_more(
-            &mut i,
-            r#"
-~> event
-  | :pressed(x), test/eager-guard(x) => {
-      selected := x
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        )
-        .unwrap_err();
-
-        assert_eq!(error.kind_name(), "ActivationPatternGuardMustBePure");
-        assert_eq!(compile_calls.load(Ordering::SeqCst), 0);
-        assert_eq!(i.symbols().borrow().snapshot(), symbols);
-        assert_eq!(plan_snapshot(&i), topology);
-        assert!(i.plan().pattern_activation_registrations().is_empty());
-        assert_eq!(i.plan().activation_registration_depth(), 0);
-    }
-
-    #[test]
-    fn activation_guard_rejects_eager_nested_match_control_flow() {
-        let mut i = interpret("event := 1.0");
-        let symbols = i.symbols().borrow().snapshot();
-        let topology = plan_snapshot(&i);
-
-        let error = interpret_more(
-            &mut i,
-            r#"
-~> event
-  | x, (x? | 0.0 => false | * => true.) => {
-      selected := x
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        )
-        .unwrap_err();
-
-        assert_eq!(error.kind_name(), "ActivationPatternGuardMustBePure");
-        assert_eq!(i.symbols().borrow().snapshot(), symbols);
-        assert_eq!(plan_snapshot(&i), topology);
-    }
-
-    #[test]
-    fn activation_guard_rejects_user_function_pattern_dispatch_that_cannot_refresh_statically() {
-        let mut i = interpret(
-            r#"
-passes(value<f64>) => <bool>
-  | 0.0 => false
-  | * => true.
-
-event := 1.0
-"#,
-        );
-        let topology = plan_snapshot(&i);
-
-        let error = interpret_more(
-            &mut i,
-            r#"
-~> event
-  | x, passes(x) => {
-      selected := x
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        )
-        .unwrap_err();
-
-        assert_eq!(error.kind_name(), "ActivationPatternGuardMustBePure");
-        assert_eq!(plan_snapshot(&i), topology);
     }
 
     #[test]
@@ -2033,486 +771,6 @@ text := "abc"
         assert_eq!(committed_capture_value(&i, 0, 0), Value::F64(Ref::new(3.0)));
     }
 
-    #[cfg(all(feature = "matrix", feature = "f64"))]
-    #[test]
-    fn activation_pattern_expression_uses_outer_symbol_when_capture_name_collides() {
-        let mut i = interpret(
-            r#"
-x := 9.0
-event := [1.0 10.0]
-
-~> event
-  | [x, x + 1.0] => {
-      selected := x
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let outer_cell = root_cell(&i, "x");
-        let activation = registration(&i);
-        let proposed_cell = proposed_capture_value(&i, 0, 0).reactive_root_cell_ids()[0];
-        let committed_cell = activation.arms[0].captures[0].cell;
-        assert_ne!(proposed_cell, committed_cell);
-        assert_ne!(outer_cell, proposed_cell);
-        assert_ne!(outer_cell, committed_cell);
-        let topology = plan_snapshot(&i);
-
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-
-        assert_dispatch_turn(&i, &topology, &outcome, 0, 1.0);
-        assert_eq!(symbol(&i, "x"), Value::F64(Ref::new(9.0)));
-        assert_eq!(proposed_capture_value(&i, 0, 0), Value::F64(Ref::new(1.0)));
-        assert_eq!(committed_capture_value(&i, 0, 0), Value::F64(Ref::new(1.0)));
-        assert_eq!(
-            proposed_capture_value(&i, 0, 0).reactive_root_cell_ids()[0],
-            proposed_cell
-        );
-        assert_eq!(registration(&i).arms[0].captures[0].cell, committed_cell);
-    }
-
-    #[cfg(all(feature = "matrix", feature = "f64"))]
-    #[test]
-    fn activation_guard_capture_shadows_outer_name_while_pattern_expression_keeps_outer_name() {
-        let mut i = interpret(
-            r#"
-x := 9.0
-event := [1.0 10.0]
-
-~> event
-  | [x, x + 1.0], x < 2.0 => {
-      selected := x + 0.0
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let outer_cell = root_cell(&i, "x");
-        let activation = registration(&i);
-        let proposed_cell = proposed_capture_value(&i, 0, 0).reactive_root_cell_ids()[0];
-        let committed_cell = activation.arms[0].captures[0].cell;
-        assert_ne!(outer_cell, proposed_cell);
-        assert_ne!(outer_cell, committed_cell);
-        assert_ne!(proposed_cell, committed_cell);
-        let guard = activation.arms[0].guard.as_ref().unwrap();
-        {
-            let plan = i.plan();
-            let plan = plan.borrow();
-            let guard_pulse_cell = plan.node(guard.match_gate_node).unwrap().outputs[0];
-            for node in &plan.nodes[guard.guard_node_start..guard.guard_node_end] {
-                assert!(node.inputs.iter().any(|dependency| {
-                    dependency.cell == guard_pulse_cell
-                        && dependency.kind == ReactiveDependencyKind::Reactive
-                }));
-                assert!(node.inputs.iter().any(|dependency| {
-                    dependency.cell == proposed_cell
-                        && dependency.kind == ReactiveDependencyKind::Sampled
-                }));
-                assert!(!node
-                    .inputs
-                    .iter()
-                    .any(|dependency| dependency.cell == committed_cell));
-            }
-        }
-        let topology = plan_snapshot(&i);
-
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-
-        assert_dispatch_turn(&i, &topology, &outcome, 0, 1.0);
-        assert_eq!(symbol(&i, "x"), Value::F64(Ref::new(9.0)));
-        assert_eq!(proposed_capture_value(&i, 0, 0), Value::F64(Ref::new(1.0)));
-        assert_eq!(committed_capture_value(&i, 0, 0), Value::F64(Ref::new(1.0)));
-    }
-
-    #[cfg(all(feature = "matrix", feature = "f64"))]
-    #[test]
-    fn activation_guard_composite_rest_proposal_commits_only_when_the_guard_passes() {
-        let mut i = interpret(
-            r#"
-event := [1.0 2.0 3.0]
-~> event
-  | [head | rest], head < rest[1] && rest[2] > rest[1] => {
-      selected := head + 0.0
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let topology = plan_snapshot(&i);
-
-        set_f64_matrix_event(&i, vec![4.0, 5.0, 6.0]);
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_eq!(selected_arm_index(&registration(&i), &outcome), 0);
-        assert_eq!(body_output_f64(&i, 0), 4.0);
-        assert_eq!(plan_snapshot(&i), topology);
-        let committed_before = committed_capture_value(&i, 0, 1);
-        let pulse_before = arm_pulse_generation(&i, 0);
-
-        set_f64_matrix_event(&i, vec![7.0, 6.0, 5.0]);
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 1, -1.0);
-        let Value::MatrixF64(proposed_rest) = proposed_capture_value(&i, 0, 1) else {
-            panic!("proposed rest is not an f64 matrix")
-        };
-        assert_eq!(proposed_rest.as_vec(), vec![6.0, 5.0]);
-        assert_eq!(committed_capture_value(&i, 0, 1), committed_before);
-        assert_eq!(arm_pulse_generation(&i, 0), pulse_before);
-        assert_eq!(body_output_f64(&i, 0), 4.0);
-        assert_eq!(plan_snapshot(&i), topology);
-    }
-
-    #[cfg(all(feature = "matrix", feature = "f64"))]
-    #[test]
-    fn activation_array_pattern_samples_expression_only_on_trigger() {
-        let mut i = interpret(
-            r#"
-event := [1.0 2.0 1.0]
-threshold := 2.0
-~> event
-  | [x, threshold + 0.0, x] => {
-      selected := x + 100.0
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let threshold_cell = root_cell(&i, "threshold");
-        let topology = plan_snapshot(&i);
-        let registration = registration(&i);
-
-        let Value::F64(threshold) = symbol(&i, "threshold") else {
-            panic!("threshold is not f64")
-        };
-        *threshold.borrow_mut() = 3.0;
-        let dependency_turn = i.advance_reactive_turn(&[threshold_cell]).unwrap();
-        let dependency_nodes = turn_executed_nodes(&dependency_turn);
-        assert!(!dependency_nodes.contains(&registration.scope_pulse_node));
-        assert!(!dependency_nodes.contains(&registration.selector_node));
-        for arm in &registration.arms {
-            assert!(!dependency_nodes.contains(&arm.matcher_node));
-            assert!(!dependency_nodes.contains(&arm.finalizer_node));
-            assert!(!dependency_nodes.contains(&arm.gate_node));
-        }
-
-        set_f64_matrix_event(&i, vec![4.0, 3.0, 4.0]);
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 0, 104.0);
-
-        set_f64_matrix_event(&i, vec![4.0, 3.0, 5.0]);
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 1, -1.0);
-    }
-
-    #[cfg(all(feature = "matrix", feature = "f64"))]
-    #[test]
-    fn activation_pattern_samples_current_user_function_output_on_trigger() {
-        let mut i = interpret(
-            r#"
-sample(value<f64>) => <f64>
-  | value + 0.0.
-
-event := [1.0 2.0 1.0]
-threshold := 2.0
-~> event
-  | [x, sample(threshold), x] => {
-      selected := x + 100.0
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let threshold_cell = root_cell(&i, "threshold");
-        let topology = plan_snapshot(&i);
-        let activation = registration(&i);
-
-        let Value::F64(threshold) = symbol(&i, "threshold") else {
-            panic!("threshold is not f64")
-        };
-        *threshold.borrow_mut() = 3.0;
-        let dependency_turn = i.advance_reactive_turn(&[threshold_cell]).unwrap();
-        let dependency_nodes = turn_executed_nodes(&dependency_turn);
-        assert!(!dependency_nodes.contains(&activation.scope_pulse_node));
-        assert!(!dependency_nodes.contains(&activation.selector_node));
-        for arm in &activation.arms {
-            assert!(!dependency_nodes.contains(&arm.matcher_node));
-            assert!(!dependency_nodes.contains(&arm.finalizer_node));
-            assert!(!dependency_nodes.contains(&arm.gate_node));
-        }
-
-        set_f64_matrix_event(&i, vec![4.0, 3.0, 4.0]);
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 0, 104.0);
-    }
-
-    #[cfg(all(feature = "matrix", feature = "f64"))]
-    #[test]
-    fn activation_array_pattern_supports_prefix_suffix_and_anonymous_spread() {
-        let mut i = interpret(
-            r#"
-event := [1.0 2.0 3.0 1.0]
-~> event
-  | [x, ..., x] => {
-      selected := x + 10.0
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let topology = plan_snapshot(&i);
-
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 0, 11.0);
-
-        set_f64_matrix_event(&i, vec![1.0, 2.0, 3.0, 4.0]);
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 1, -1.0);
-    }
-
-    #[cfg(all(feature = "matrix", feature = "f64"))]
-    #[test]
-    fn activation_array_rest_segment_accepts_nested_array_pattern() {
-        let mut i = interpret(
-            r#"
-event := [1.0 2.0 3.0 4.0]
-~> event
-  | [head | [second, ..., last]] => {
-      selected := head * 100.0 + second * 10.0 + last
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let topology = plan_snapshot(&i);
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 0, 124.0);
-    }
-
-    #[cfg(feature = "u64")]
-    #[test]
-    fn activation_typed_literal_pattern_uses_shared_value_matching() {
-        let mut i = interpret(
-            r#"
-event := 1u64
-~> event
-  | 1u64 => {
-      selected := 1.0
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let topology = plan_snapshot(&i);
-        let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &outcome, 0, 1.0);
-    }
-
-    #[test]
-    fn activation_whole_composite_capture_is_stable_and_visible_to_the_body() {
-        let mut i = interpret(
-            r#"
-event := (1.0, 2.0)
-~> event
-  | whole => {
-      selected := whole
-    }
-  | * => {
-      selected := (-1.0, -1.0)
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let activation = registration(&i);
-        let capture = &activation.arms[0].captures[0];
-        assert_eq!(capture.kind, ValueKind::Tuple(vec![ValueKind::F64, ValueKind::F64]));
-        let body_inputs = i.plan().borrow().nodes
-            [activation.arms[0].body_node_start..activation.arms[0].body_node_end]
-            .iter()
-            .flat_map(|node| node.inputs.iter().map(|dependency| dependency.cell))
-            .collect::<Vec<_>>();
-        assert!(
-            body_inputs.contains(&capture.cell),
-            "capture cell {:?} is absent from body inputs {:?}",
-            capture.cell,
-            body_inputs
-        );
-        let topology = plan_snapshot(&i);
-        for values in [[3.0, 4.0], [5.0, 6.0]] {
-            set_tuple_event(
-                &i,
-                values
-                    .into_iter()
-                    .map(|value| Value::F64(Ref::new(value)))
-                    .collect(),
-            );
-            let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-            assert_eq!(selected_arm_index(&activation, &outcome), 0);
-            assert_eq!(
-                body_output(&i, 0),
-                Value::Tuple(Ref::new(MechTuple::from_vec(
-                    values
-                        .into_iter()
-                        .map(|value| Value::F64(Ref::new(value)))
-                        .collect(),
-                )))
-            );
-            assert_eq!(registration(&i).arms[0].captures[0].cell, capture.cell);
-            assert_eq!(plan_snapshot(&i), topology);
-        }
-    }
-
-    #[test]
-    fn activation_whole_tuple_capture_keeps_element_access_attached() {
-        let mut i = interpret(
-            r#"
-event := (1.0, 2.0)
-~> event
-  | whole => {
-      selected := whole.1 * 10.0 + whole.2
-    }
-  | * => {
-      selected := -1.0
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let topology = plan_snapshot(&i);
-        for (values, expected) in [([3.0, 4.0], 34.0), ([5.0, 6.0], 56.0)] {
-            set_tuple_event(
-                &i,
-                values
-                    .into_iter()
-                    .map(|value| Value::F64(Ref::new(value)))
-                    .collect(),
-            );
-            let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-            assert_dispatch_turn(&i, &topology, &outcome, 0, expected);
-        }
-    }
-
-    #[cfg(all(feature = "matrix", feature = "f64"))]
-    #[test]
-    fn activation_array_rest_capture_preserves_kind_payload_and_identity() {
-        let mut i = interpret(
-            r#"
-event := [1.0 2.0 3.0 4.0 5.0]
-~> event
-  | [head | rest] => {
-      selected := rest
-    }
-  | * => {
-      selected := [-1.0]
-    }
-"#,
-        );
-        let trigger = root_cell(&i, "event");
-        let activation = registration(&i);
-        let rest_capture = &activation.arms[0].captures[1];
-        assert_eq!(
-            rest_capture.kind,
-            ValueKind::Matrix(Box::new(ValueKind::F64), Vec::new())
-        );
-        assert!(
-            i.plan().borrow().nodes
-                [activation.arms[0].body_node_start..activation.arms[0].body_node_end]
-                .iter()
-                .any(|node| node
-                    .inputs
-                    .iter()
-                    .any(|dependency| dependency.cell == rest_capture.cell))
-        );
-        let topology = plan_snapshot(&i);
-        let Value::MatrixF64(event) = symbol(&i, "event") else {
-            panic!("event is not an f64 matrix")
-        };
-        for values in [
-            vec![10.0, 20.0, 30.0, 40.0, 50.0],
-            vec![11.0, 21.0, 31.0, 41.0, 51.0, 61.0],
-        ] {
-            let source = Matrix::from_vec(values.clone(), 1, values.len());
-            assert!(event.replace_payload_from(&source));
-            let outcome = i.advance_reactive_turn(&[trigger]).unwrap();
-            assert_eq!(selected_arm_index(&activation, &outcome), 0);
-            let Value::MatrixF64(rest) = body_output(&i, 0) else {
-                panic!("rest output is not an f64 matrix")
-            };
-            assert_eq!(rest.shape(), vec![1, values.len() - 1]);
-            assert_eq!(rest.as_vec(), values[1..]);
-            assert_eq!(
-                registration(&i).arms[0].captures[1].cell,
-                rest_capture.cell
-            );
-            assert_eq!(plan_snapshot(&i), topology);
-        }
-    }
-    #[test]
-    fn activation_pattern_repeated_capture_kind_mismatch_uses_canonical_error() {
-        let mut i = interpret("event := (1.0, \"one\")");
-        let topology = plan_snapshot(&i);
-        let error = interpret_more(
-            &mut i,
-            "~> event\n  | (x, x) => {
-      selected := x
-    }\n  | * => {
-      selected := 0.0
-    }",
-        )
-        .unwrap_err();
-        assert_eq!(error.kind_name(), "PatternCompileError");
-        assert_eq!(plan_snapshot(&i), topology);
-        assert!(!i.symbols().borrow().contains(hash_str("x")));
-        assert!(!i.symbols().borrow().contains(hash_str("selected")));
-    }
-
-    #[test]
-    fn activation_pattern_capture_does_not_leak() {
-        let (mut i, trigger, topology) = tuple_fixture(FLAT_TUPLE_ACTIVATION);
-        for name in ["x", "y", "selected"] {
-            assert!(!i.symbols().borrow().contains(hash_str(name)));
-        }
-        set_tuple_event(&i, vec![Value::F64(Ref::new(3.)), Value::F64(Ref::new(4.))]);
-        let o = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &o, 0, 34.);
-    }
-    #[test]
-    fn activation_pattern_capture_shadows_and_restores_outer_symbol() {
-        let mut i = interpret("event := (1.0, 2.0)\nx := 99.0");
-        let outer = symbol_ref(&i, "x");
-        let address = outer.addr();
-        interpret_more(
-            &mut i,
-            "~> event\n  | (x, y) => {
-      selected := x + y
-    }\n  | * => {
-      selected := -1.0
-    }",
-        )
-        .unwrap();
-        assert_eq!(*symbol(&i, "x").as_f64().unwrap().borrow(), 99.);
-        assert_eq!(symbol_ref(&i, "x").addr(), address);
-        assert!(!i.symbols().borrow().contains(hash_str("y")));
-        assert!(!i.symbols().borrow().contains(hash_str("selected")));
-        let topology = plan_snapshot(&i);
-        let trigger = root_cell(&i, "event");
-        set_tuple_event(&i, vec![Value::F64(Ref::new(3.)), Value::F64(Ref::new(4.))]);
-        let o = i.advance_reactive_turn(&[trigger]).unwrap();
-        assert_dispatch_turn(&i, &topology, &o, 0, 7.);
-    }
     fn failed_elaboration_fixture() -> (
         Interpreter,
         SymbolTableSnapshot,
